backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
celery
redis
passlib[bcrypt]
python-jose[cryptography]

# AI & Image Processing
torch
transformers
ftfy
regex
tqdm
Pillow
git+https://github.com/openai/CLIP.git
requests 